delete popup added in popUps and works

+++ src/app/reply/reply.component.ts
@@ -11,8 +11,9 @@ import {
 import { FormControl } from "@angular/forms";
 import { KweeksService } from "../services/kweeks.service";
 import { Overlay } from "@angular/cdk/overlay";
-import { LikesRekweeksListComponent } from '../likes-rekweeks-list/likes-rekweeks-list.component';
-import { NewKweekComponent } from '../new-kweek/new-kweek.component';
+import { LikesRekweeksListComponent } from "../likes-rekweeks-list/likes-rekweeks-list.component";
+import { NewKweekComponent } from "../new-kweek/new-kweek.component";
+import { ConfirmDeleteComponent } from "../confirm-delete/confirm-delete.component";
 @Component({
   selector: "app-reply",
   templateUrl: "./reply.component.html",
@@ -88,7 +89,10 @@ export class ReplyComponent implements OnInit {
     const dialogConfig = new MatDialogConfig();
     dialogConfig.width = "520px";
     dialogConfig.autoFocus = false;
-    const dialogLikersRekweekersRef = this.dialog.open(LikesRekweeksListComponent, dialogConfig);
+    const dialogLikersRekweekersRef = this.dialog.open(
+      LikesRekweeksListComponent,
+      dialogConfig
+    );
     dialogLikersRekweekersRef.componentInstance.clickedKweek = kweek;
     dialogLikersRekweekersRef.componentInstance.likers = true;
   }
@@ -102,7 +106,10 @@ export class ReplyComponent implements OnInit {
     const dialogConfig = new MatDialogConfig();
     dialogConfig.width = "520px";
     dialogConfig.autoFocus = false;
-    const dialogLikersRekweekersRef = this.dialog.open(LikesRekweeksListComponent, dialogConfig);
+    const dialogLikersRekweekersRef = this.dialog.open(
+      LikesRekweeksListComponent,
+      dialogConfig
+    );
     dialogLikersRekweekersRef.componentInstance.clickedKweek = kweek;
     dialogLikersRekweekersRef.componentInstance.likers = false;
   }
@@ -164,13 +171,25 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   deleteRoot_ClickedKweek(kweek: Kweek): void {
-    if (!this.busyRequest) {
-      this.busyRequest = true;
-      this.kweekService.deleteKweek(kweek.id).subscribe(() => {
-        this.dialogRef.close();
-        this.busyRequest = false;
-      });
-    }
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.width = "520px";
+    dialogConfig.autoFocus = false;
+    const confirmDeleteRef = this.dialog.open(
+      ConfirmDeleteComponent,
+      dialogConfig
+    );
+    confirmDeleteRef.componentInstance.clickedKweek = kweek;
+    confirmDeleteRef.afterClosed().subscribe(res => {
+      if (res) {
+        if (!this.busyRequest) {
+          this.busyRequest = true;
+          this.kweekService.deleteKweek(kweek.id).subscribe(() => {
+            this.dialogRef.close();
+            this.busyRequest = false;
+          });
+        }
+      }
+    });
   }
 
   /**
@@ -179,22 +198,33 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   deleteReply(kweek: Kweek): void {
-    if (!this.busyRequest) {
-      this.busyRequest = true;
-      this.kweekService.deleteKweek(kweek.id).subscribe(() => {
-        const indexToDelete = this.replies.indexOf(kweek);
-        this.replies.splice(indexToDelete, 1);
-        this.busyRequest = false;
-      });
-    }
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.width = "520px";
+    dialogConfig.autoFocus = false;
+    const confirmDeleteRef = this.dialog.open(
+      ConfirmDeleteComponent,
+      dialogConfig
+    );
+    confirmDeleteRef.componentInstance.clickedKweek = kweek;
+    confirmDeleteRef.afterClosed().subscribe(res => {
+      if (res) {
+        if (!this.busyRequest) {
+          this.busyRequest = true;
+          this.kweekService.deleteKweek(kweek.id).subscribe(() => {
+            const indexToDelete = this.replies.indexOf(kweek);
+            this.replies.splice(indexToDelete, 1);
+            this.busyRequest = false;
+          });
+        }
+      }
+    });
   }
 
   Reply(kweek: Kweek): void {
     const dialogRef = this.dialog.open(NewKweekComponent, {
-      panelClass: 'kweekBox'
+      panelClass: "kweekBox"
     });
     dialogRef.componentInstance.kweek = kweek;
     dialogRef.componentInstance.reply = true;
   }
-
 }
